refactor: factor out string types into subpackage
