+++ src/bisque/element/sentinels.py
@@ -1,0 +1,5 @@
+__all__ = ["DEFAULT_TYPES_SENTINEL"]
+
+
+class DEFAULT_TYPES_SENTINEL:
+    """Replaces a plain object."""
