+++ src/bisque/element/tag_core/main.py
@@ -7,8 +7,24 @@ from bisque.formatter import Formatter
 
 from ...typing.tabulation import BaseTypeTable
 from ..encodings import DEFAULT_OUTPUT_ENCODING
+from ..sentinels import DEFAULT_TYPES_SENTINEL
 from .page_element import BasePageElement
 from .soup_strainer import BaseSoupStrainer
+from .string_types import (
+    BaseCData,
+    BaseComment,
+    BaseDeclaration,
+    BaseDoctype,
+    BaseNavigableString,
+    BasePreformattedString,
+    BaseProcessingInstruction,
+    BaseRubyParenthesisString,
+    BaseRubyTextString,
+    BaseScript,
+    BaseStylesheet,
+    BaseTemplateString,
+    BaseXMLProcessingInstruction,
+)
 from .tag import BaseTag
 
 __all__ = [
@@ -81,138 +97,15 @@ class PageElement(BasePageElement, TabulatedType):
 # Section 2: Text strings (13 classes)
 
 
-class NavigableString(str, PageElement, TabulatedType):
+class NavigableString(BaseNavigableString, PageElement, TabulatedType):
     """A Python Unicode string that is part of a parse tree.
 
     When Bisque parses the markup <b>penguin</b>, it will
     create a NavigableString for the string "penguin".
     """
 
-    PREFIX = ""
-    SUFFIX = ""
 
-    def __new__(cls, value):
-        """Create a new NavigableString.
-
-        When unpickling a NavigableString, this method is called with
-        the string in DEFAULT_OUTPUT_ENCODING. That encoding needs to be
-        passed in to the superclass's __new__ or the superclass won't know
-        how to handle non-ASCII characters.
-        """
-        if isinstance(value, str):
-            u = str.__new__(cls, value)
-        else:
-            u = str.__new__(cls, value, DEFAULT_OUTPUT_ENCODING)
-        u.setup()
-        return u
-
-    def __deepcopy__(self, memo, recursive=False):
-        """A copy of a NavigableString has the same contents and class
-        as the original, but it is not connected to the parse tree.
-
-        :param recursive: This parameter is ignored; it's only defined
-           so that NavigableString.__deepcopy__ implements the same
-           signature as Tag.__deepcopy__.
-        """
-        return type(self)(self)
-
-    def __copy__(self):
-        """A copy of a NavigableString can only be a deep copy, because
-        only one PageElement can occupy a given place in a parse tree.
-        """
-        return self.__deepcopy__({})
-
-    def __getnewargs__(self):
-        return (str(self),)
-
-    def __getattr__(self, attr):
-        """text.string gives you text. This is for backwards
-        compatibility for Navigable*String, but for CData* it lets you
-        get the string without the CData wrapper."""
-        if attr == "string":
-            return self
-        else:
-            raise AttributeError(
-                "'{}' object has no attribute '{}'".format(
-                    self.__class__.__name__,
-                    attr,
-                ),
-            )
-
-    def output_ready(self, formatter="minimal"):
-        """Run the string through the provided formatter.
-
-        :param formatter: A Formatter object, or a string naming one of the standard formatters.
-        """
-        output = self.format_string(self, formatter)
-        return self.PREFIX + output + self.SUFFIX
-
-    @property
-    def name(self):
-        """Since a NavigableString is not a Tag, it has no .name.
-
-        This property is implemented so that code like this doesn't crash
-        when run on a mixture of Tag and NavigableString objects:
-            [x.name for x in tag.children]
-        """
-        return None
-
-    @name.setter
-    def name(self, name):
-        """Prevent NavigableString.name from ever being set."""
-        raise AttributeError("A NavigableString cannot be given a name.")
-
-    def _all_strings(self, strip=False, types=PageElement.default):
-        """Yield all strings of certain classes, possibly stripping them.
-
-        This makes it easy for NavigableString to implement methods
-        like get_text() as conveniences, creating a consistent
-        text-extraction API across all PageElements.
-
-        :param strip: If True, all strings will be stripped before being
-            yielded.
-
-        :param types: A tuple of NavigableString subclasses. If this
-            NavigableString isn't one of those subclasses, the
-            sequence will be empty. By default, the subclasses
-            considered are NavigableString and CData objects. That
-            means no comments, processing instructions, etc.
-
-        :yield: A sequence that either contains this string, or is empty.
-
-        """
-        if types is self.default:
-            # This is kept in Tag because it's full of subclasses of
-            # this class, which aren't defined until later in the file.
-            types = Tag.DEFAULT_INTERESTING_STRING_TYPES
-
-        # Do nothing if the caller is looking for specific types of
-        # string, and we're of a different type.
-        #
-        # We check specific types instead of using isinstance(self,
-        # types) because all of these classes subclass
-        # NavigableString. Anyone who's using this feature probably
-        # wants generic NavigableStrings but not other stuff.
-        my_type = type(self)
-        if types is not None:
-            if isinstance(types, type):
-                # Looking for a single type.
-                if my_type is not types:
-                    return
-            elif my_type not in types:
-                # Looking for one of a list of types.
-                return
-
-        value = self
-        if strip:
-            value = value.strip()
-        if len(value) > 0:
-            yield value
-
-    strings = property(_all_strings)
-
-
-class PreformattedString(NavigableString, TabulatedType):
+class PreformattedString(BasePreformattedString, NavigableString, TabulatedType):
     """A NavigableString not subject to the normal formatting rules.
 
     This is an abstract class used for special kinds of strings such
@@ -220,93 +113,40 @@ class PreformattedString(NavigableString, TabulatedType):
     class).
     """
 
-    PREFIX = ""
-    SUFFIX = ""
 
-    def output_ready(self, formatter=None):
-        """Make this string ready for output by adding any subclass-specific
-            prefix or suffix.
-
-        :param formatter: A Formatter object, or a string naming one
-            of the standard formatters. The string will be passed into the
-            Formatter, but only to trigger any side effects: the return
-            value is ignored.
-
-        :return: The string, with any subclass-specific prefix and
-           suffix added on.
-        """
-        if formatter is not None:
-            # this used to assign to an unused var named "ignore"
-            self.format_string(self, formatter)
-        return self.PREFIX + self + self.SUFFIX
-
-
-class CData(PreformattedString, TabulatedType):
+class CData(BaseCData, PreformattedString, TabulatedType):
     """A CDATA block."""
 
-    PREFIX = "<![CDATA["
-    SUFFIX = "]]>"
 
-
-class ProcessingInstruction(PreformattedString, TabulatedType):
+class ProcessingInstruction(
+    BaseProcessingInstruction,
+    PreformattedString,
+    TabulatedType,
+):
     """A SGML processing instruction."""
 
-    PREFIX = "<?"
-    SUFFIX = ">"
 
-
-class XMLProcessingInstruction(ProcessingInstruction, TabulatedType):
+class XMLProcessingInstruction(
+    BaseXMLProcessingInstruction,
+    ProcessingInstruction,
+    TabulatedType,
+):
     """An XML processing instruction."""
 
-    PREFIX = "<?"
-    SUFFIX = "?>"
 
-
-class Comment(PreformattedString, TabulatedType):
+class Comment(BaseComment, PreformattedString, TabulatedType):
     """An HTML or XML comment."""
 
-    PREFIX = "<!--"
-    SUFFIX = "-->"
 
-
-class Declaration(PreformattedString, TabulatedType):
+class Declaration(BaseDeclaration, PreformattedString, TabulatedType):
     """An XML declaration."""
 
-    PREFIX = "<?"
-    SUFFIX = "?>"
 
-
-class Doctype(PreformattedString, TabulatedType):
+class Doctype(BaseDoctype, PreformattedString, TabulatedType):
     """A document type declaration."""
 
-    @classmethod
-    def for_name_and_ids(cls, name, pub_id, system_id):
-        """Generate an appropriate document type declaration for a given
-        public ID and system ID.
 
-        :param name: The name of the document's root element, e.g. 'html'.
-        :param pub_id: The Formal Public Identifier for this document type,
-            e.g. '-//W3C//DTD XHTML 1.1//EN'
-        :param system_id: The system identifier for this document type,
-            e.g. 'http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd'
-
-        :return: A Doctype.
-        """
-        value = name or ""
-        if pub_id is not None:
-            value += ' PUBLIC "%s"' % pub_id
-            if system_id is not None:
-                value += ' "%s"' % system_id
-        elif system_id is not None:
-            value += ' SYSTEM "%s"' % system_id
-
-        return Doctype(value)
-
-    PREFIX = "<!DOCTYPE "
-    SUFFIX = ">\n"
-
-
-class Stylesheet(NavigableString, TabulatedType):
+class Stylesheet(BaseStylesheet, NavigableString, TabulatedType):
     """A NavigableString representing an stylesheet (probably
     CSS).
 
@@ -314,7 +154,7 @@ class Stylesheet(NavigableString, TabulatedType):
     """
 
 
-class Script(NavigableString, TabulatedType):
+class Script(BaseScript, NavigableString, TabulatedType):
     """A NavigableString representing an executable script (probably
     Javascript).
 
@@ -322,7 +162,7 @@ class Script(NavigableString, TabulatedType):
     """
 
 
-class TemplateString(NavigableString, TabulatedType):
+class TemplateString(BaseTemplateString, NavigableString, TabulatedType):
     """A NavigableString representing a string found inside an HTML
     template embedded in a larger document.
 
@@ -330,7 +170,7 @@ class TemplateString(NavigableString, TabulatedType):
     """
 
 
-class RubyTextString(NavigableString, TabulatedType):
+class RubyTextString(BaseRubyTextString, NavigableString, TabulatedType):
     """A NavigableString representing the contents of the <rt> HTML
     element.
 
@@ -341,7 +181,7 @@ class RubyTextString(NavigableString, TabulatedType):
     """
 
 
-class RubyParenthesisString(NavigableString, TabulatedType):
+class RubyParenthesisString(BaseRubyParenthesisString, NavigableString, TabulatedType):
     """A NavigableString representing the contents of the <rp> HTML
     element.
 
@@ -419,7 +259,7 @@ class Tag(BaseTag, PageElement, TabulatedType):
 
     DEFAULT_INTERESTING_STRING_TYPES = (NavigableString, CData)
 
-    def _all_strings(self, strip=False, types=PageElement.default):
+    def _all_strings(self, strip=False, types=DEFAULT_TYPES_SENTINEL):
         """Yield all strings of certain classes, possibly stripping them.
 
         :param strip: If True, all strings will be stripped before being
@@ -436,7 +276,7 @@ class Tag(BaseTag, PageElement, TabulatedType):
         :yield: A sequence of strings.
 
         """
-        if types is self.default:
+        if types is DEFAULT_TYPES_SENTINEL:
             types = self.interesting_string_types
 
         for descendant in self.descendants:
